bot/pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Telegram bot for lab assistant"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=20.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"